pyproject.toml
[tool.poetry]
name = "griptape-trade-school"
version = "0.1.0"
description = ""
authors = ["Jason Schleifer <jason.schleifer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.26"
mkdocs-glightbox = "^0.4.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
python-dotenv = "^1.0.1"
pytest-xdist = "^3.3.1"
griptape = {version="^0.28.2", extras=["all"]}
pyright = "^1.1.372"
typos = "^1.23.2"
ruff = "^0.6.0"

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
